+++ src-tauri/src/prefs.rs
@@ -6,7 +6,7 @@ use std::{
     sync::Mutex,
 };
 
-use anyhow::{anyhow, ensure, Result};
+use anyhow::{anyhow, ensure, Context, Result};
 use log::{debug, info, warn};
 use serde::{Deserialize, Serialize};
 use tauri::{AppHandle, Manager};
@@ -286,17 +286,21 @@ impl Prefs {
 
     pub fn create(app: &AppHandle) -> Result<Self> {
         let path = Self::path();
-        fs::create_dir_all(path.parent().unwrap())?;
+        fs::create_dir_all(path.parent().unwrap())
+            .context("failed to create settings directory")?;
+
+        info!("loading settings from {}", path.display());
 
         let is_first_run = !path.exists();
         let prefs = match is_first_run {
             true => {
+                info!("no settings file found, creating new default");
+
                 let prefs = Prefs {
                     is_first_run,
                     ..Default::default()
                 };
 
-                prefs.save()?;
                 prefs
             }
             false => {