+++ CHANGELOG.md
@@ -4,6 +4,23 @@
 
 ### Added
 
+- More logging and error context, especially during the startup process
+
+### Fixed
+
+- Modpack export causing the app to freeze on games without a "Modpack" category
+
+## 1.0.1 (2024-11-09)
+
+### Fixed
+
+- Importing profiles with local mods (they are ignored now)
+- r2modman/Thunderstore Mod Manager profile transfer for Risk Of Rain 2
+
+## 1.0.0 (2024-11-09)
+
+### Added
+
 - `Open game directory` button to File menu
 
 ## 0.9.2 (2024-10-29)